feat: separe component Home
