+++ src/modules/comic/composables/index.ts
@@ -1,1 +1,1 @@
-export * from './useComic'
+export * from './useHomeView'
