+++ src/modules/comic/views/HomeView.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
-import { useComic } from '../composables'
+import { useHomeView } from '../composables'
 
-const { loader, comic, setStar, starClass } = useComic()
+const { HomeInfo, HomeActions, comic, loader } = useHomeView()
 </script>
 
 <template>
@@ -11,26 +11,9 @@ const { loader, comic, setStar, starClass } = useComic()
     </template>
 
     <section class="home-section" v-else>
-      <h1 class="home-section__h1">{{ comic.getComic.title.toUpperCase() }}</h1>
-      <img class="home-section__img" :src="comic.getComic.img" :alt="comic.getComic.alt" />
+      <HomeInfo />
 
-      <div class="home-section__stars">
-        <h3>Puntuacion:</h3>
-        <p>
-          <template v-for="star in 5" :key="star">
-            <i
-              @click="setStar(star)"
-              class="home-section__stars__icon"
-              :class="starClass(star)"
-            ></i>
-          </template>
-        </p>
-      </div>
-
-      <article class="home-section__btns">
-        <button v-if="comic.getPage > 1" @click="comic.setPage(comic.getPage - 1)">prev</button>
-        <button @click="comic.setPage(comic.getPage + 1)">next</button>
-      </article>
+      <HomeActions />
     </section>
   </main>
 </template>
